Fix as discussed on the developer list: Check for the availability of mailpreferences before using them.
git-svn-id: http://svn.egroupware.org/egroupware/trunk@32953 923e3c2e-ae13-0410-ba18-ced7211e33de

+++ felamimail/inc/class.bofelamimail.inc.php
@@ -218,17 +218,19 @@ class bofelamimail
 		*/
 		function addAccount($_hookValues)
 		{
-			$icServer = $this->mailPreferences->getIncomingServer(0);
-			if(is_a($icServer,'defaultimap')) {
-				// if not connected, try opening an admin connection
-				if (!$icServer->_connected) $this->openConnection(0,true);
-				$icServer->addAccount($_hookValues);
-				if ($icServer->_connected) $this->closeConnection(); // close connection afterwards
-			}
+			if ($this->mailPreferences) {
+				$icServer = $this->mailPreferences->getIncomingServer(0);
+				if(is_a($icServer,'defaultimap')) {
+					// if not connected, try opening an admin connection
+					if (!$icServer->_connected) $this->openConnection(0,true);
+					$icServer->addAccount($_hookValues);
+					if ($icServer->_connected) $this->closeConnection(); // close connection afterwards
+				}
 
-			$ogServer = $this->mailPreferences->getOutgoingServer(0);
-			if(is_a($ogServer,'defaultsmtp')) {
-				$ogServer->addAccount($_hookValues);
+				$ogServer = $this->mailPreferences->getOutgoingServer(0);
+				if(is_a($ogServer,'defaultsmtp')) {
+					$ogServer->addAccount($_hookValues);
+				}
 			}
 		}
 
@@ -524,17 +526,19 @@ class bofelamimail
 
 		function deleteAccount($_hookValues)
 		{
-			$icServer = $this->mailPreferences->getIncomingServer(0);
-			if(is_a($icServer,'defaultimap')) {
-				//try to connect with admin rights, when not connected
-				if (!$icServer->_connected) $this->openConnection(0,true);
-				$icServer->deleteAccount($_hookValues);
-				if ($icServer->_connected) $this->closeConnection(); // close connection
-			}
+			if ($this->mailPreferences) {
+				$icServer = $this->mailPreferences->getIncomingServer(0);
+				if(is_a($icServer,'defaultimap')) {
+					//try to connect with admin rights, when not connected
+					if (!$icServer->_connected) $this->openConnection(0,true);
+					$icServer->deleteAccount($_hookValues);
+					if ($icServer->_connected) $this->closeConnection(); // close connection
+				}
 
-			$ogServer = $this->mailPreferences->getOutgoingServer(0);
-			if(is_a($ogServer,'defaultsmtp')) {
-				$ogServer->deleteAccount($_hookValues);
+				$ogServer = $this->mailPreferences->getOutgoingServer(0);
+				if(is_a($ogServer,'defaultsmtp')) {
+					$ogServer->deleteAccount($_hookValues);
+				}
 			}
 		}
